app/requirements.txt
fastapi
uvicorn
requests
aiohttp
aiofiles
boto3
pandas
numpy
sqlalchemy
asyncpg
python-dotenv
docker
openai
bs4
lxml
alembic
selenium
webdriver-manager
investpy
pyjwt
google_auth_oauthlib
google-api-python-client
jinja2